feat: 選手追加フィールドを共通コンポーネントに切り出し
