+++ Localizable.xcstrings
@@ -1445,6 +1445,9 @@
         }
       }
     },
+    "追加" : {
+
+    },
     "選手一覧" : {
       "localizations" : {
         "en" : {
